+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `sncast get spec-version` command that returns the version of the Starknet JSON-RPC specification used by the node
 - `--no-abi` flag for `declare`, `declare-from`, and `deploy`, erasing the ABI before class declaration.
+- `sncast get tx-receipt` (alias: `get transaction-receipt`) command that returns the receipt of a transaction
 
 #### Fixed